+++ raphtory-benchmark/src/main.rs
@@ -191,10 +191,7 @@ fn main() {
     // Out neighbours of all nodes with time
     now = Instant::now();
     let nodes = &g.nodes();
-    let _out_neighbours = nodes
-        .iter()
-        .map(|v| v.out_neighbours())
-        .collect::<Vec<_>>();
+    let _out_neighbours = nodes.iter().map(|v| v.out_neighbours()).collect::<Vec<_>>();
     println!("Out neighbours: {} seconds", now.elapsed().as_secs_f64());
 
     // page rank with time
