+++ raphtory-graphql/src/data.rs
@@ -1,6 +1,6 @@
 use parking_lot::RwLock;
 #[cfg(feature = "storage")]
-use raphtory::disk_graph::graph_impl::DiskGraphStorage;
+use raphtory::disk_graph::DiskGraphStorage;
 use raphtory::{
     core::Prop,
     db::api::view::MaterializedGraph,
