+++ raphtory/src/db/api/storage/storage_ops/edge_filter.rs
@@ -1,7 +1,6 @@
 use crate::{
     core::entities::LayerIds,
     db::api::{storage::edges::edge_ref::EdgeStorageRef, view::internal::EdgeFilterOps},
-    disk_graph::DiskGraph,
 };
 
 use super::GraphStorage;
