+++ raphtory/src/disk_graph/graph_impl/mod.rs
@@ -1,28 +1,9 @@
+use std::path::Path;
+
 use crate::{
-    core::{
-        entities::{
-            properties::{graph_meta::GraphMeta, props::Meta},
-            LayerIds,
-        },
-        utils::errors::GraphError,
-    },
-    disk_graph::{graph_impl::prop_conversion::make_node_properties_from_graph, Error},
-    prelude::{Graph, Layer},
+    disk_graph::{DiskGraphError, DiskGraphStorage},
+    prelude::Graph,
 };
-use pometry_storage::{
-    disk_hmap::DiskHashMap, graph::TemporalGraph, graph_fragment::TempColGraphFragment,
-    load::ExternalEdgeList, RAError,
-};
-use raphtory_api::core::entities::edges::edge_ref::EdgeRef;
-use rayon::prelude::*;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
-use std::{
-    fmt::{Display, Formatter},
-    path::{Path, PathBuf},
-    sync::Arc,
-};
-};
-use crate::disk_graph::DiskGraphError;
 
 mod edge_storage_ops;
 mod interop;
@@ -39,55 +20,6 @@ pub struct ParquetLayerCols<'a> {
     pub time_col: &'a str,
 }
 
-#[derive(Clone, Debug)]
-pub struct DiskGraphStorage {
-    pub(crate) inner: Arc<TemporalGraph>,
-    node_meta: Arc<Meta>,
-    edge_meta: Arc<Meta>,
-    graph_props: Arc<GraphMeta>,
-    graph_dir: PathBuf,
-}
-
-impl Serialize for DiskGraphStorage {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let path = self.graph_dir.clone();
-        path.serialize(serializer)
-    }
-}
-
-impl<'de> Deserialize<'de> for DiskGraphStorage {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let path = PathBuf::deserialize(deserializer)?;
-        let graph_result = DiskGraphStorage::load_from_dir(&path).map_err(|err| {
-            serde::de::Error::custom(format!("Failed to load Diskgraph: {:?}", err))
-        })?;
-        Ok(graph_result)
-    }
-}
-
-impl Display for DiskGraphStorage {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "Diskgraph(num_nodes={}, num_temporal_edges={}",
-            self.inner.num_nodes(),
-            self.inner.count_temporal_edges()
-        )
-    }
-}
-
-impl AsRef<TemporalGraph> for DiskGraphStorage {
-    fn as_ref(&self) -> &TemporalGraph {
-        &self.inner
-    }
-}
-
 impl Graph {
     pub fn persist_as_disk_graph(
         &self,
@@ -97,338 +29,27 @@ impl Graph {
     }
 }
 
-fn get_valid_layers(graph: &Arc<TemporalGraph>) -> Vec<String> {
-    graph.layer_names().into_iter().map(|x| x.clone()).collect()
-}
-
-impl DiskGraphStorage {
-    pub fn valid_layer_ids_from_names(&self, key: Layer) -> LayerIds {
-        match key {
-            Layer::All | Layer::Default => LayerIds::One(0), // FIXME: need to handle all correctly
-            Layer::One(name) => {
-                let name = name.as_ref();
-                self.inner
-                    .layer_names()
-                    .iter()
-                    .enumerate()
-                    .find(move |(_, ref n)| n == &name)
-                    .map(|(i, _)| LayerIds::One(i))
-                    .unwrap_or(LayerIds::None)
-            }
-            _ => todo!("Layer ids for multiple names not implemented for Diskgraph"),
-        }
-    }
-
-    pub fn layer_ids_from_names(&self, key: Layer) -> Result<LayerIds, GraphError> {
-        match key {
-            Layer::All => Ok(LayerIds::All),
-            Layer::Default => Ok(LayerIds::One(0)),
-            Layer::One(name) => {
-                let id = self.inner.find_layer_id(&name).ok_or_else(|| {
-                    GraphError::invalid_layer(name.to_string(), get_valid_layers(&self.inner))
-                })?;
-                Ok(LayerIds::One(id))
-            }
-            Layer::None => Ok(LayerIds::None),
-            Layer::Multiple(names) => {
-                let ids = names
-                    .iter()
-                    .map(|name| {
-                        self.inner.find_layer_id(name).ok_or_else(|| {
-                            GraphError::invalid_layer(
-                                name.to_string(),
-                                get_valid_layers(&self.inner),
-                            )
-                        })
-                    })
-                    .collect::<Result<Vec<_>, _>>()?;
-                Ok(LayerIds::Multiple(ids.into()))
-            }
-        }
-    }
-
-    pub fn into_graph(self) -> Graph {
-        Graph::from_internal_graph(&crate::db::api::storage::storage_ops::GraphStorage::Disk(
-            Arc::new(self),
-        ))
-    }
-
-    pub(crate) fn core_temporal_edge_prop_ids(
-        &self,
-        e: EdgeRef,
-        layer_ids: &LayerIds,
-    ) -> Box<dyn Iterator<Item = usize> + '_> {
-        let layer_id = match layer_ids.constrain_from_edge(e) {
-            LayerIds::One(id) => id,
-            _ => panic!("Only one layer is supported"),
-        };
-        Box::new(1..self.inner.edges_data_type(layer_id).len())
-    }
-
-    pub fn layer_from_ids(&self, layer_ids: &LayerIds) -> Option<usize> {
-        match layer_ids {
-            LayerIds::One(layer_id) => Some(*layer_id),
-            LayerIds::None => None,
-            LayerIds::All => match self.inner.layers().len() {
-                0 => None,
-                1 => Some(0),
-                _ => todo!("multilayer edge views not yet supported in Diskgraph"),
-            },
-            LayerIds::Multiple(ids) => match ids.len() {
-                0 => None,
-                1 => Some(ids[0]),
-                _ => todo!("multilayer edge views not yet supported in Diskgraph"),
-            },
-        }
-    }
-
-    pub fn make_simple_graph(
-        graph_dir: impl AsRef<Path>,
-        edges: &[(u64, u64, i64, f64)],
-        chunk_size: usize,
-        t_props_chunk_size: usize,
-    ) -> DiskGraphStorage {
-        // unzip into 4 vectors
-        let (src, (dst, (time, weight))): (Vec<_>, (Vec<_>, (Vec<_>, Vec<_>))) = edges
-            .iter()
-            .map(|(a, b, c, d)| (*a, (*b, (*c, *d))))
-            .unzip();
-
-        let edge_lists = vec![StructArray::new(
-            DataType::Struct(vec![
-                Field::new("src", DataType::UInt64, false),
-                Field::new("dst", DataType::UInt64, false),
-                Field::new("time", DataType::Int64, false),
-                Field::new("weight", DataType::Float64, false),
-            ]),
-            vec![
-                PrimitiveArray::from_vec(src).boxed(),
-                PrimitiveArray::from_vec(dst).boxed(),
-                PrimitiveArray::from_vec(time).boxed(),
-                PrimitiveArray::from_vec(weight).boxed(),
-            ],
-            None,
-        )];
-        DiskGraphStorage::load_from_edge_lists(
-            &edge_lists,
-            chunk_size,
-            t_props_chunk_size,
-            graph_dir.as_ref(),
-            0,
-            1,
-            2,
-        )
-        .expect("failed to create graph")
-    }
-
-    fn new(inner_graph: TemporalGraph, graph_dir: PathBuf) -> Self {
-        let node_meta = Meta::new();
-        let mut edge_meta = Meta::new();
-        let graph_meta = GraphMeta::new();
-
-        for node_type in inner_graph.node_types().into_iter().flatten() {
-            if let Some(node_type) = node_type {
-                node_meta.get_or_create_node_type_id(node_type);
-            } else {
-                panic!("Node types cannot be null");
-            }
-        }
-
-        for layer in inner_graph.layers() {
-            let edge_props_fields = layer.edges_data_type();
-
-            for (id, field) in edge_props_fields.iter().enumerate() {
-                let prop_name = &field.name;
-                let data_type = field.data_type();
-
-                let resolved_id = edge_meta
-                    .resolve_prop_id(prop_name, data_type.into(), false)
-                    .expect("Arrow data types should without failing");
-                if id != resolved_id {
-                    println!("Warning: Layers with different edge properties are not supported by the high-level apis on top of the disk_graph graph yet, edge properties will not be available to high-level apis");
-                    edge_meta = Meta::new();
-                    break;
-                }
-            }
-        }
-
-        for l_name in inner_graph.layer_names() {
-            edge_meta.layer_meta().get_or_create_id(l_name);
-        }
-
-        if let Some(props) = &inner_graph.node_properties().const_props {
-            let node_const_props_fields = props.prop_dtypes();
-            for field in node_const_props_fields {
-                node_meta
-                    .resolve_prop_id(&field.name, field.data_type().into(), true)
-                    .expect("Initial resolve should not fail");
-            }
-        }
-
-        if let Some(props) = &inner_graph.node_properties().temporal_props {
-            let node_temporal_props_fields = props.prop_dtypes();
-            for field in node_temporal_props_fields {
-                node_meta
-                    .resolve_prop_id(&field.name, field.data_type().into(), false)
-                    .expect("Initial resolve should not fail");
-            }
-        }
-
-        Self {
-            inner: Arc::new(inner_graph),
-            node_meta: Arc::new(node_meta),
-            edge_meta: Arc::new(edge_meta),
-            graph_props: Arc::new(graph_meta),
-            graph_dir,
-        }
-    }
-
-    pub fn from_graph(graph: &Graph, graph_dir: impl AsRef<Path>) -> Result<Self, Error> {
-        let inner_graph = TemporalGraph::from_graph(graph, graph_dir.as_ref(), || {
-            make_node_properties_from_graph(graph, graph_dir.as_ref())
-        })?;
-        Ok(Self::new(inner_graph, graph_dir.as_ref().to_path_buf()))
-    }
-
-    pub fn load_from_edge_lists(
-        edge_list: &[StructArray],
-        chunk_size: usize,
-        t_props_chunk_size: usize,
-        graph_dir: impl AsRef<Path> + Sync,
-        src_col_idx: usize,
-        dst_col_idx: usize,
-        time_col_idx: usize,
-    ) -> Result<Self, RAError> {
-        let path = graph_dir.as_ref().to_path_buf();
-        let inner = TemporalGraph::from_sorted_edge_list(
-            graph_dir,
-            src_col_idx,
-            dst_col_idx,
-            time_col_idx,
-            chunk_size,
-            t_props_chunk_size,
-            edge_list,
-        )?;
-        Ok(Self::new(inner, path))
-    }
-
-    pub fn load_from_dir(graph_dir: impl AsRef<Path>) -> Result<DiskGraphStorage, RAError> {
-        let path = graph_dir.as_ref().to_path_buf();
-        let inner = TemporalGraph::new(graph_dir)?;
-        Ok(Self::new(inner, path))
-    }
-
-    pub fn load_from_parquets<P: AsRef<Path>>(
-        graph_dir: P,
-        layer_parquet_cols: Vec<ParquetLayerCols>,
-        node_properties: Option<P>,
-        chunk_size: usize,
-        t_props_chunk_size: usize,
-        read_chunk_size: Option<usize>,
-        concurrent_files: Option<usize>,
-        num_threads: usize,
-        node_type_col: Option<&str>,
-    ) -> Result<DiskGraphStorage, RAError> {
-        let layered_edge_list: Vec<ExternalEdgeList<&Path>> = layer_parquet_cols
-            .iter()
-            .map(
-                |ParquetLayerCols {
-                     parquet_dir,
-                     layer,
-                     src_col,
-                     dst_col,
-                     time_col,
-                 }| {
-                    ExternalEdgeList::new(layer, parquet_dir.as_ref(), src_col, dst_col, time_col)
-                        .expect("Failed to load events")
-                },
-            )
-            .collect::<Vec<_>>();
-
-        let t_graph = TemporalGraph::from_parquets(
-            num_threads,
-            chunk_size,
-            t_props_chunk_size,
-            read_chunk_size,
-            concurrent_files,
-            graph_dir.as_ref(),
-            layered_edge_list,
-            node_properties.as_ref().map(|p| p.as_ref()),
-            node_type_col,
-        )?;
-        Ok(Self::new(t_graph, graph_dir.as_ref().to_path_buf()))
-    }
-
-    pub fn filtered_layers_par<'a>(
-        &'a self,
-        layer_ids: &'a LayerIds,
-    ) -> impl ParallelIterator<Item = &'a TempColGraphFragment> + 'a {
-        self.inner
-            .layers()
-            .par_iter()
-            .enumerate()
-            .filter(|(l_id, _)| layer_ids.contains(l_id))
-            .map(|(_, layer)| layer)
-    }
-
-    pub fn filtered_layers_iter<'a>(
-        &'a self,
-        layer_ids: &'a LayerIds,
-    ) -> impl Iterator<Item = &'a TempColGraphFragment> + 'a {
-        self.inner
-            .layers()
-            .iter()
-            .enumerate()
-            .filter(|(l_id, _)| layer_ids.contains(l_id))
-            .map(|(_, layer)| layer)
-    }
-
-    pub fn from_layer(layer: TempColGraphFragment) -> Self {
-        let path = layer.graph_dir().to_path_buf();
-        let global_ordering = layer.nodes_storage().gids().clone();
-
-        let global_order = DiskHashMap::from_sorted_dedup(global_ordering.clone())
-            .expect("Failed to create global order");
-
-        let inner = TemporalGraph::new_from_layers(
-            global_ordering,
-            Arc::new(global_order),
-            vec![layer],
-            vec!["_default".to_string()],
-        );
-        Self::new(inner, path)
-    }
-
-    pub fn node_meta(&self) -> &Meta {
-        &self.node_meta
-    }
-
-    pub fn edge_meta(&self) -> &Meta {
-        &self.edge_meta
-    }
-
-    pub fn graph_meta(&self) -> &GraphMeta {
-        &self.graph_props
-    }
-}
-
 #[cfg(test)]
 mod test {
-    use super::{DiskGraphStorage, ParquetLayerCols};
+    use std::{
+        path::{Path, PathBuf},
+        sync::Arc,
+    };
+
+    use itertools::Itertools;
+    use proptest::{prelude::*, sample::size_range};
+    use rayon::prelude::*;
+    use tempfile::TempDir;
+
+    use pometry_storage::{graph::TemporalGraph, properties::Properties};
+
     use crate::{
         db::api::{storage::storage_ops::GraphStorage, view::StaticGraphViewOps},
         disk_graph::Time,
         prelude::*,
     };
-    use itertools::Itertools;
-    use pometry_storage::{graph::TemporalGraph, properties::Properties};
-    use proptest::{prelude::*, sample::size_range};
-    use rayon::prelude::*;
-    use std::{
-        path::{Path, PathBuf},
-        sync::Arc,
-    };
-    use tempfile::TempDir;
+
+    use super::{DiskGraphStorage, ParquetLayerCols};
 
     fn make_simple_graph(graph_dir: impl AsRef<Path>, edges: &[(u64, u64, i64, f64)]) -> Graph {
         let storage = DiskGraphStorage::make_simple_graph(graph_dir, edges, 1000, 1000);
@@ -998,16 +619,19 @@ mod test {
 #[cfg(feature = "storage")]
 #[cfg(test)]
 mod storage_tests {
+    use std::collections::BTreeSet;
+
+    use itertools::Itertools;
+    use proptest::prelude::*;
+    use tempfile::TempDir;
+
+    use raphtory_api::core::storage::arc_str::OptionAsStr;
+
     use crate::{
         core::Prop,
         db::graph::graph::assert_graph_equal,
         prelude::{AdditionOps, Graph, GraphViewOps, NodeViewOps, NO_PROPS, *},
     };
-    use itertools::Itertools;
-    use proptest::prelude::*;
-    use raphtory_api::core::storage::arc_str::OptionAsStr;
-    use std::collections::BTreeSet;
-    use tempfile::TempDir;
 
     #[test]
     fn test_merge() {
@@ -1049,7 +673,10 @@ mod storage_tests {
         let g1_a = g1.persist_as_disk_graph(&g1_dir).unwrap();
         let g2_a = g2.persist_as_disk_graph(&g2_dir).unwrap();
 
-        let gm = g1_a.merge_by_sorted_gids(&g2_a, &gm_dir).unwrap();
+        let gm = g1_a
+            .merge_by_sorted_gids(&g2_a, &gm_dir)
+            .unwrap()
+            .into_graph();
 
         let n0 = gm.node(0).unwrap();
         assert_eq!(
@@ -1083,15 +710,27 @@ mod storage_tests {
         assert_eq!(n2.node_type().as_str(), Some("3")); // right has priority
 
         assert_eq!(
-            gm.default_layer().edges().id().collect::<Vec<_>>(),
+            gm.default_layer()
+                .edges()
+                .id()
+                .filter_map(|(a, b)| a.as_u64().zip(b.as_u64()))
+                .collect::<Vec<_>>(),
             [(0, 1), (1, 2)]
         );
         assert_eq!(
-            gm.valid_layers("1").edges().id().collect::<Vec<_>>(),
+            gm.valid_layers("1")
+                .edges()
+                .id()
+                .filter_map(|(a, b)| a.as_u64().zip(b.as_u64()))
+                .collect::<Vec<_>>(),
             [(0, 1)]
         );
         assert_eq!(
-            gm.valid_layers("2").edges().id().collect::<Vec<_>>(),
+            gm.valid_layers("2")
+                .edges()
+                .id()
+                .filter_map(|(a, b)| a.as_u64().zip(b.as_u64()))
+                .collect::<Vec<_>>(),
             [(0, 1)]
         );
     }
@@ -1128,7 +767,7 @@ mod storage_tests {
         let merged_g_disk = left_g_disk
             .merge_by_sorted_gids(&right_g_disk, &merged_dir)
             .unwrap();
-        assert_graph_equal(&merged_g_disk, &merged_g_expected)
+        assert_graph_equal(&merged_g_disk.into_graph(), &merged_g_expected)
     }
 
     #[test]
