--- raphtory/src/disk_graph/query/executors/mod.rs
@@ -1,1 +1,0 @@
-pub mod rayon2;
